requirements.txt
ultralytics>=8.0.0
opencv-python
pyzbar
pillow
numpy
albumentations
tqdm
matplotlib
